Move handleMessage to handlers.js
